pyproject.toml
[tool.poetry]
name = "rag-chatbot"
version = "0.1.0"
description = "Enterprise RAG Chatbot — Multi-cloud (AWS + Azure) retrieval-augmented generation service with FastAPI"
authors = ["Ketan Sahu <Ketan.Sahu@odido.nl>"]
readme = "README.md"
license = "MIT"
packages = [{ include = "src" }]

[tool.poetry.dependencies]
python = "^3.12"

# --- Web Framework ---
fastapi = "^0.115.0"
uvicorn = { version = "^0.32.0", extras = ["standard"] }
python-multipart = "^0.0.12"
jinja2 = "^3.1.4"
sse-starlette = "^2.1.0"

# --- Pydantic ---
pydantic = "^2.9.0"
pydantic-settings = "^2.5.0"

# --- LangChain (RAG orchestration) ---
langchain = "^0.3.0"
langchain-community = "^0.3.0"
langchain-text-splitters = "^0.3.0"

# --- AWS ---
boto3 = "^1.35.0"
langchain-aws = "^0.2.0"
opensearch-py = "^2.7.0"

# --- Azure ---
azure-identity = "^1.19.0"
azure-storage-blob = "^12.23.0"
azure-search-documents = "^11.6.0"
azure-cosmos = "^4.9.0"
openai = "^1.50.0"

# --- Document Processing ---
pypdf = "^5.0.0"
python-docx = "^1.1.0"

# --- Monitoring & Observability ---
opentelemetry-api = "^1.27.0"
opentelemetry-sdk = "^1.27.0"
opentelemetry-instrumentation-fastapi = "^0.48b0"
opentelemetry-exporter-otlp = "^1.27.0"
loguru = "^0.7.2"

# --- Utilities ---
httpx = "^0.27.0"
tenacity = "^9.0.0"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
pytest-asyncio = "^0.24.0"
pytest-cov = "^5.0.0"
httpx = "^0.27.0"
ruff = "^0.6.0"
pre-commit = "^3.8.0"
moto = {version = "^5.0.0", extras = ["s3", "dynamodb", "lambda"]}
localstack = "^3.8.0"

[tool.poetry.scripts]
start = "src.main:run"

[tool.ruff]
target-version = "py312"
line-length = 120
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP", "B", "SIM", "RUF"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-v --tb=short"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
